+++ integration_tests/src/tests/mod.rs
@@ -8,7 +8,8 @@ mod change_beneficiary_test;
 pub use change_beneficiary_test::*;
 mod change_owner_test;
 pub use change_owner_test::*;
-// mod commit_post_test;
+mod commit_post_test;
+pub use commit_post_test::*;
 // mod datacap_tests;
 // mod evm_test;
 // mod extend_sectors_test;
